Fix parsing of numbered lists with prefixed markers in SSJC
Previously, if the first item of a numbered list in a SSJC entry
was not exactly equal to the string "1.", it failed to be recognized
as a start of a new list.

This has been fixed to also recognize the first item, if the number
is prefixed with any non-numeric string.